third_party/blink/web_tests/external/wpt/audio-output/setSinkId-with-selectAudioOutput.https-expected.txt

This is a testharness.js-based test.
[FAIL] setSinkId() after selectAudioOutput()
  promise_test: Unhandled rejection with value: object "TypeError: navigator.mediaDevices.selectAudioOutput is not a function"
[FAIL] setSinkId() on video after loadeddata
  assert_not_equals: selectAudioOutput() resolved got disallowed value undefined
[FAIL] setSinkId() on video after play
  assert_not_equals: selectAudioOutput() resolved got disallowed value undefined
[FAIL] setSinkId() with deviceID from another window
  assert_not_equals: selectAudioOutput() resolved got disallowed value undefined
Harness: the test ran to completion.
